+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Known bugs:
 
 Breaking changes:
 
+- The exit code for an error of the reader module is changed from 8 to 64 and for a writer from 7 to 128. (:issue:`773`)
+
 New features and notable changes:
 
 - Add `--html-template-dir` option to use custom Jinja2 templates (:issue:`758`)
@@ -19,6 +21,7 @@ New features and notable changes:
 - Fix Dockerfile.qa to avoid uid conflicts. (:issue:`801`)
 - Pygments required >= 2.13.0. (:issue:`799`)
 - Add a second theme for HTML report inspired by GitHub. (:issue:`793`)
+- Add :option:`--fail-under-decision` and :option:`--fail-under-function` which will error under a given minimum coverage. (:issue:`773`)
 
 
 Bug fixes and small improvements:
@@ -32,6 +35,7 @@ Bug fixes and small improvements:
 - Fix symlinked root directories on Windows. (:issue:`814`)
 - Extend :option:`--gcov-ignore-errors` to be able to ignore specific gcov errors. (:issue:`787`)
 - Fix reading of choices options from configuration files (e.g. ``gcov-ignore-parse-errors``). (:issue:`816`)
+- Fix ``TypeError`` during decision analysis. (:issue:`784`)
 
 Documentation: